requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.2.3
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0
opencv-python==4.8.0.76 